removed old domain impl to add reverse engineering impl. Added the modifications to mormalize domains
