+++ src/main/java/dlt/dltbackendmaster/controller/ServiceController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import dlt.dltbackendmaster.domain.Service;
 import dlt.dltbackendmaster.domain.ServiceType;
+import dlt.dltbackendmaster.domain.Services;
 import dlt.dltbackendmaster.service.DAOService;
 
 @RestController
@@ -25,10 +25,10 @@ public class ServiceController
     }
 
     @GetMapping(produces = "application/json")
-    public ResponseEntity<List<Service>> getAll() {
+    public ResponseEntity<List<Services>> getAll() {
 
         try {
-            List<Service> services = service.getAll(Service.class);
+            List<Services> services = service.getAll(Services.class);
             return new ResponseEntity<>(services, HttpStatus.OK);
         } catch (Exception e) {
             e.printStackTrace();
@@ -37,14 +37,14 @@ public class ServiceController
     }
 
     @GetMapping(path = "/{serviceType}", produces = "application/json")
-    public ResponseEntity<List<Service>> get(@PathVariable ServiceType serviceType) {
+    public ResponseEntity<List<Services>> get(@PathVariable ServiceType serviceType) {
 
         if (serviceType == null) {
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
 
         try {
-            List<Service> services = service.GetAllEntityByNamedQuery("Service.findByServiceType", serviceType.ordinal());
+            List<Services> services = service.GetAllEntityByNamedQuery("Service.findByServiceType", serviceType.ordinal()+1);
             return new ResponseEntity<>(services, HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
