+++ src/main/java/dlt/dltbackendmaster/controller/SyncController.java
@@ -22,14 +22,14 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import dlt.dltbackendmaster.domain.Beneficiary;
-import dlt.dltbackendmaster.domain.BeneficiaryIntervention;
+import dlt.dltbackendmaster.domain.Beneficiaries;
+import dlt.dltbackendmaster.domain.BeneficiariesInterventions;
 import dlt.dltbackendmaster.domain.Locality;
 import dlt.dltbackendmaster.domain.Neighborhood;
 import dlt.dltbackendmaster.domain.Partners;
 import dlt.dltbackendmaster.domain.Profiles;
-import dlt.dltbackendmaster.domain.Service;
-import dlt.dltbackendmaster.domain.SubService;
+import dlt.dltbackendmaster.domain.Services;
+import dlt.dltbackendmaster.domain.SubServices;
 import dlt.dltbackendmaster.domain.Us;
 import dlt.dltbackendmaster.domain.Users;
 import dlt.dltbackendmaster.domain.watermelondb.BeneficiaryInterventionSyncModel;
@@ -69,20 +69,20 @@ public class SyncController {
 		List<Us> usCreated;
 		List<Us> usUpdated;
 		
-		List<Beneficiary> beneficiariesCreated;
-		List<Beneficiary> beneficiariesUpdated;
+		List<Beneficiaries> beneficiariesCreated;
+		List<Beneficiaries> beneficiariesUpdated;
         
-        List<BeneficiaryIntervention> beneficiariesInterventionsCreated;
-        List<BeneficiaryIntervention> beneficiariesInterventionsUpdated;
+        List<BeneficiariesInterventions> beneficiariesInterventionsCreated;
+        List<BeneficiariesInterventions> beneficiariesInterventionsUpdated;
         
         List<Neighborhood> neighborhoodsCreated;
         List<Neighborhood> neighborhoodUpdated;
         
-        List<Service> servicesCreated;
-        List<Service> servicesUpdated;
+        List<Services> servicesCreated;
+        List<Services> servicesUpdated;
         
-        List<SubService> subServicesCreated;
-        List<SubService> subServicesUpdated;
+        List<SubServices> subServicesCreated;
+        List<SubServices> subServicesUpdated;
 		
 		if(lastPulledAt == null || lastPulledAt.equals("null") ) {
 
@@ -105,19 +105,19 @@ public class SyncController {
 			usUpdated = new ArrayList<Us>();
 			
 			beneficiariesCreated = service.GetAllEntityByNamedQuery("Beneficiary.findAll");
-			beneficiariesUpdated = new ArrayList<Beneficiary>();
+			beneficiariesUpdated = new ArrayList<Beneficiaries>();
             
             beneficiariesInterventionsCreated = service.GetAllEntityByNamedQuery("BeneficiaryIntervention.findAll");
-            beneficiariesInterventionsUpdated = new ArrayList<BeneficiaryIntervention>();
+            beneficiariesInterventionsUpdated = new ArrayList<BeneficiariesInterventions>();
             
             neighborhoodsCreated = service.GetAllEntityByNamedQuery("Neighborhood.findAll");
             neighborhoodUpdated = new ArrayList<Neighborhood>();
             
             servicesCreated = service.GetAllEntityByNamedQuery("Service.findAll");
-            servicesUpdated = new ArrayList<Service>();
+            servicesUpdated = new ArrayList<Services>();
             
             subServicesCreated = service.GetAllEntityByNamedQuery("SubService.findAll");
-            subServicesUpdated = new ArrayList<SubService>();
+            subServicesUpdated = new ArrayList<SubServices>();
 			
 		}else {
 			Long t = Long.valueOf(lastPulledAt);
@@ -169,11 +169,11 @@ public class SyncController {
 			SyncObject<Partners> partnersSO = new SyncObject<Partners>(partnersCreated, partnersUpdated, listDeleted);
 			SyncObject<Profiles> profilesSO = new SyncObject<Profiles>(profilesCreated, profilesUpdated, listDeleted);
 			SyncObject<Us> usSO = new SyncObject<Us>(usCreated, usUpdated, listDeleted);
-			SyncObject<Beneficiary> beneficiarySO = new SyncObject<Beneficiary>(beneficiariesCreated, beneficiariesUpdated, listDeleted);
-            SyncObject<BeneficiaryIntervention> beneficiaryInterventionSO = new SyncObject<BeneficiaryIntervention>(beneficiariesInterventionsCreated, beneficiariesInterventionsUpdated, listDeleted);
+			SyncObject<Beneficiaries> beneficiarySO = new SyncObject<Beneficiaries>(beneficiariesCreated, beneficiariesUpdated, listDeleted);
+            SyncObject<BeneficiariesInterventions> beneficiaryInterventionSO = new SyncObject<BeneficiariesInterventions>(beneficiariesInterventionsCreated, beneficiariesInterventionsUpdated, listDeleted);
             SyncObject<Neighborhood> neighborhoodSO = new SyncObject<Neighborhood>(neighborhoodsCreated, neighborhoodUpdated, listDeleted);
-            SyncObject<Service> serviceSO = new SyncObject<Service>(servicesCreated, servicesUpdated, listDeleted);
-            SyncObject<SubService> subServiceSO = new SyncObject<SubService>(subServicesCreated, subServicesUpdated, listDeleted);
+            SyncObject<Services> serviceSO = new SyncObject<Services>(servicesCreated, servicesUpdated, listDeleted);
+            SyncObject<SubServices> subServiceSO = new SyncObject<SubServices>(subServicesCreated, subServicesUpdated, listDeleted);
 
 			//String object = SyncSerializer.createUsersSyncObject(usersCreated, usersUpdated, new ArrayList<Integer>());
 
@@ -228,9 +228,8 @@ public class SyncController {
 			    
 			    for (BeneficiarySyncModel created : createdList) {
                     if(created.getOnline_id() == null) {
-                        Beneficiary beneficiary = new Beneficiary(created, lastPulledAt);
-                        beneficiary.getCreatedBy().setId(user.getId());
-                        //beneficiary.setCreatedBy(user.getId());
+                    	Beneficiaries beneficiary = new Beneficiaries(created, lastPulledAt);
+                    	beneficiary.setCreatedBy(user.getId());
                         service.Save(beneficiary);
                     }
                 }
@@ -240,7 +239,7 @@ public class SyncController {
                 
                 for (BeneficiaryInterventionSyncModel created : createdList) {
                     if(created.getOnline_id() == null) {
-                        BeneficiaryIntervention intervention = new BeneficiaryIntervention(created, lastPulledAt);
+                    	BeneficiariesInterventions intervention = new BeneficiariesInterventions(created, lastPulledAt);
                         intervention.setCreatedBy(user.getId());
                         service.Save(intervention);
                     }
@@ -273,15 +272,15 @@ public class SyncController {
                 for (BeneficiarySyncModel updated : updatedList) {
                     
                     if(updated.getOnline_id() == null) {
-                        Beneficiary beneficiary = new Beneficiary(updated, lastPulledAt);
-                        beneficiary.getCreatedBy().setId(user.getId());
-                        //beneficiary.setCreatedBy(user.getId());
+                    	Beneficiaries beneficiary = new Beneficiaries(updated, lastPulledAt);
+                        //beneficiary.getCreatedBy().setId(user.getId());
+                        beneficiary.setCreatedBy(user.getId());
                         service.Save(beneficiary);
                         
                     } else {
-                        Beneficiary beneficiary = service.find(Beneficiary.class, updated.getOnline_id());
-                        beneficiary.getUpdatedBy().setId(user.getId());
-                        //beneficiary.setUpdatedBy(user.getId());
+                    	Beneficiaries beneficiary = service.find(Beneficiaries.class, updated.getOnline_id());
+                        //beneficiary.getUpdatedBy().setId(user.getId());
+                        beneficiary.setUpdatedBy(user.getId());
                         beneficiary.update(updated, lastPulledAt);
                         service.update(beneficiary);
                     } 
@@ -293,15 +292,15 @@ public class SyncController {
                 for (BeneficiaryInterventionSyncModel updated : updatedList) {
                     
                     if(updated.getOnline_id() == null) {
-                        BeneficiaryIntervention intervention = new BeneficiaryIntervention(updated, lastPulledAt);
+                    	BeneficiariesInterventions intervention = new BeneficiariesInterventions(updated, lastPulledAt);
                         intervention.setCreatedBy(user.getId());
                         service.Save(intervention);
                         
                     } else {
-                        BeneficiaryIntervention intervention = service.find(BeneficiaryIntervention.class, updated.getOnline_id());
-                        intervention.setUpdatedBy(user.getId());
-                        intervention.update(updated, lastPulledAt);
-                        service.update(intervention);
+                        //BeneficiaryIntervention intervention = service.find(BeneficiaryIntervention.class, updated.getOnline_id());
+                        //intervention.setUpdatedBy(user.getId());
+                       // intervention.update(updated, lastPulledAt);
+                        //service.update(intervention);
                     } 
                 }
             }
