+++ src/main/java/dlt/dltbackendmaster/controller/UserController.java
@@ -35,6 +35,7 @@ public class UserController
             List<Users> users = service.getAll(Users.class);
             return new ResponseEntity<>(users, HttpStatus.OK);
         } catch (Exception e) {
+        	System.out.println(e.getMessage());
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
