+++ src/main/java/dlt/dltbackendmaster/domain/Beneficiaries.java
@@ -1,6 +1,8 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -12,17 +14,37 @@ import static javax.persistence.GenerationType.IDENTITY;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.NamedQueries;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import javax.persistence.UniqueConstraint;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+import dlt.dltbackendmaster.domain.watermelondb.BeneficiarySyncModel;
+import dlt.dltbackendmaster.serializers.LocalitySerializer;
+import dlt.dltbackendmaster.serializers.NeighborhoodSerializer;
+import dlt.dltbackendmaster.serializers.PartnersSerializer;
+import dlt.dltbackendmaster.serializers.UsSerializer;
+
 /**
  * Beneficiaries generated by hbm2java
  */
 @Entity
 @Table(name = "beneficiaries", catalog = "dreams_db", uniqueConstraints = @UniqueConstraint(columnNames = "nui"))
+@NamedQueries({ 
+	@NamedQuery(name = "Beneficiaries.findAll", query = "SELECT b FROM Beneficiaries b"),
+    @NamedQuery(name = "Beneficiaries.findByDateCreated",
+                query = "select b from Beneficiaries b where b.dateUpdated is null and b.dateCreated > :lastpulledat"),
+    @NamedQuery(name = "Beneficiaries.findByDateUpdated",
+                query = "select b from Beneficiaries b where (b.dateUpdated >= :lastpulledat) or (b.dateUpdated >= :lastpulledat and b.dateCreated = b.dateUpdated)") })
 public class Beneficiaries implements java.io.Serializable {
 
 	private Integer id;
@@ -161,6 +183,103 @@ public class Beneficiaries implements java.io.Serializable {
 		this.beneficiariesInterventionses = beneficiariesInterventionses;
 		this.referenceses = referenceses;
 	}
+	
+	public Beneficiaries(BeneficiarySyncModel model, String timestamp) {
+        Long t = Long.valueOf(timestamp);
+        Date regDate = new Date(t);
+        this.nui = model.getNui();
+        this.surname = model.getSurname();
+        this.nickName = model.getNick_name();
+        this.partners = new Partners(model.getOrganization_id());
+        this.dateOfBirth = model.getDate_of_birth();
+        this.gender = model.getGender();
+        this.address = model.getAddress();
+        this.phoneNumber = model.getPhone_number();
+        this.EMail = model.getE_mail();
+        this.via = model.getVia();
+        this.partnerId = model.getPartner_id();
+        this.entryPoint = model.getEntry_point();
+        this.neighborhood = new Neighborhood(model.getNeighbourhood_id());
+        this.us = new Us(model.getUs_id()); 
+        this.offlineId = model.getId();
+        this.dateCreated = regDate;
+        this.dateUpdated = regDate;
+        this.vbltLivesWith = model.getVblt_lives_with();
+        this.vbltIsOrphan = (byte)model.getVblt_is_orphan();
+        this.vbltIsStudent = (byte)model.getVblt_is_student();
+        this.vbltSchoolGrade = model.getVblt_school_grade();
+        this.vbltSchoolName = model.getVblt_school_name();
+        this.vbltIsDeficient = (byte)model.getVblt_is_deficient();
+        this.vbltDeficiencyType = model.getVblt_deficiency_type();
+        this.vbltMarriedBefore = (byte)model.getVblt_married_before();
+        this.vbltPregnantBefore = (byte)model.getVblt_pregnant_before();
+        this.vbltChildren = (byte)model.getVblt_children();
+        this.vbltPregnantOrBreastfeeding = (byte)model.getVblt_pregnant_or_breastfeeding();
+        this.vbltIsEmployed = model.getVblt_is_employed();
+        this.vbltTestedHiv = model.getVblt_tested_hiv();
+        this.vbltSexuallyActive = (byte)model.getVblt_sexually_active();
+        this.vbltMultiplePartners = (byte)model.getVblt_multiple_partners();
+        this.vbltIsMigrant = (byte)model.getVblt_is_migrant();
+        this.vbltTraffickingVictim = (byte)model.getVblt_trafficking_victim();
+        this.vbltSexualExploitation = (byte)model.getVblt_sexual_exploitation();
+        this.vbltSexploitationTime = model.getVblt_sexploitation_time();
+        this.vbltVbgVictim = (byte)model.getVblt_vbg_victim();
+        this.vbltVbgType = model.getVblt_vbg_type();
+        this.vbltVbgTime = model.getVblt_vbg_time();
+        this.vbltAlcoholDrugsUse = (byte)model.getVblt_alcohol_drugs_use();
+        this.vbltStiHistory = (byte)model.getVblt_sti_history();
+        this.vbltSexWorker = (byte)model.getVblt_sex_worker();
+        this.vbltHouseSustainer = (byte)model.getVblt_house_sustainer();
+        
+    }
+	
+	public Beneficiaries(BeneficiarySyncModel model) {
+    	this.nui = model.getNui();
+        this.surname = model.getSurname();
+        this.nickName = model.getNick_name();
+        this.partners = new Partners(model.getOrganization_id());
+        this.dateOfBirth = model.getDate_of_birth();
+        this.gender = model.getGender();
+        this.address = model.getAddress();
+        this.phoneNumber = model.getPhone_number();
+        this.EMail = model.getE_mail();
+        this.via = model.getVia();
+        this.partnerId = model.getPartner_id();
+        this.entryPoint = model.getEntry_point();
+        this.neighborhood = new Neighborhood(model.getNeighbourhood_id());
+        this.us = new Us(model.getUs_id());
+        this.offlineId = model.getId();
+        this.vbltLivesWith = model.getVblt_lives_with();
+        this.vbltIsOrphan = (byte)model.getVblt_is_orphan();
+        this.vbltIsStudent = (byte)model.getVblt_is_student();
+        this.vbltSchoolGrade = model.getVblt_school_grade();
+        this.vbltSchoolName = model.getVblt_school_name();
+        this.vbltIsDeficient = (byte)model.getVblt_is_deficient();
+        this.vbltDeficiencyType = model.getVblt_deficiency_type();
+        this.vbltMarriedBefore = (byte)model.getVblt_married_before();
+        this.vbltPregnantBefore = (byte)model.getVblt_pregnant_before();
+        this.vbltChildren = (byte)model.getVblt_children();
+        this.vbltPregnantOrBreastfeeding = (byte)model.getVblt_pregnant_or_breastfeeding();
+        this.vbltIsEmployed = model.getVblt_is_employed();
+        this.vbltTestedHiv = model.getVblt_tested_hiv();
+        this.vbltSexuallyActive = (byte)model.getVblt_sexually_active();
+        this.vbltMultiplePartners = (byte)model.getVblt_multiple_partners();
+        this.vbltIsMigrant = (byte)model.getVblt_is_migrant();
+        this.vbltTraffickingVictim = (byte)model.getVblt_trafficking_victim();
+        this.vbltSexualExploitation = (byte)model.getVblt_sexual_exploitation();
+        this.vbltSexploitationTime = model.getVblt_sexploitation_time();
+        this.vbltVbgVictim = (byte)model.getVblt_vbg_victim();
+        this.vbltVbgType = model.getVblt_vbg_type();
+        this.vbltVbgTime = model.getVblt_vbg_time();
+        this.vbltAlcoholDrugsUse = (byte)model.getVblt_alcohol_drugs_use();
+        this.vbltStiHistory = (byte)model.getVblt_sti_history();
+        this.vbltSexWorker = (byte)model.getVblt_sex_worker();
+        this.vbltHouseSustainer = (byte)model.getVblt_house_sustainer();
+    }
+	
+	public Beneficiaries(Integer id) {
+        this.id = id;
+    }
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
@@ -174,8 +293,10 @@ public class Beneficiaries implements java.io.Serializable {
 		this.id = id;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "neighbourhood_id", nullable = false)
+	@JsonProperty("neighbourhood")
+    @JsonSerialize(using = NeighborhoodSerializer.class)
 	public Neighborhood getNeighborhood() {
 		return this.neighborhood;
 	}
@@ -184,8 +305,10 @@ public class Beneficiaries implements java.io.Serializable {
 		this.neighborhood = neighborhood;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "partner")
+	@JsonProperty("partner")
+    @JsonSerialize(using = PartnersSerializer.class)
 	public Partners getPartners() {
 		return this.partners;
 	}
@@ -194,8 +317,10 @@ public class Beneficiaries implements java.io.Serializable {
 		this.partners = partners;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "us_id", nullable = false)
+	@JsonProperty("us")
+    @JsonSerialize(using = UsSerializer.class)
 	public Us getUs() {
 		return this.us;
 	}
@@ -612,7 +737,8 @@ public class Beneficiaries implements java.io.Serializable {
 		this.offlineId = offlineId;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "beneficiaries")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "beneficiaries")
 	public Set<VulnerabilityHistory> getVulnerabilityHistories() {
 		return this.vulnerabilityHistories;
 	}
@@ -621,7 +747,8 @@ public class Beneficiaries implements java.io.Serializable {
 		this.vulnerabilityHistories = vulnerabilityHistories;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "beneficiaries")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "beneficiaries")
 	public Set<BeneficiariesInterventions> getBeneficiariesInterventionses() {
 		return this.beneficiariesInterventionses;
 	}
@@ -630,7 +757,8 @@ public class Beneficiaries implements java.io.Serializable {
 		this.beneficiariesInterventionses = beneficiariesInterventionses;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "beneficiaries")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "beneficiaries")
 	public Set<References> getReferenceses() {
 		return this.referenceses;
 	}
@@ -639,4 +767,112 @@ public class Beneficiaries implements java.io.Serializable {
 		this.referenceses = referenceses;
 	}
 
+	public ObjectNode toObjectNode() {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode beneficiary = mapper.createObjectNode();
+        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
+        if (offlineId != null) {
+            beneficiary.put("id", offlineId);
+        } else {
+            beneficiary.put("id", id);
+        }
+
+        if (dateUpdated == null || dateUpdated.after(dateCreated)) {
+            beneficiary.put("id", id);
+            beneficiary.put("nui", nui);
+            beneficiary.put("name", name);
+            beneficiary.put("surname", surname);
+            beneficiary.put("nickname", nickName);
+            beneficiary.put("organization_id", partners.getId());
+            beneficiary.put("date_of_birth", dateFormat.format(dateOfBirth));
+            beneficiary.put("gender", String.valueOf(gender));
+            beneficiary.put("adress", address);
+            beneficiary.put("phone_number", phoneNumber);
+            beneficiary.put("email", EMail);
+            beneficiary.put("via", via);
+            beneficiary.put("partner_id", partnerId == null ? null : 1); //TODO
+            beneficiary.put("entry_point", entryPoint);
+            beneficiary.put("neighborhood_id", neighborhood.getId());
+            beneficiary.put("us_id", us.getId());
+            beneficiary.put("online_id", id.toString()); // flag to control if entity is synchronized with the backend
+            beneficiary.put("vblt_lives_with", vbltLivesWith);
+            beneficiary.put("vblt_is_orphan", vbltIsOrphan);
+            beneficiary.put("vblt_is_student", vbltIsStudent);
+            beneficiary.put("vblt_school_grade", vbltSchoolGrade);
+            beneficiary.put("vblt_school_name", vbltSchoolName);
+            beneficiary.put("vblt_is_deficient", vbltIsDeficient);
+            beneficiary.put("vblt_deficiency_type", vbltDeficiencyType);
+            beneficiary.put("vblt_married_before", vbltMarriedBefore);
+            beneficiary.put("vblt_pregnant_before", vbltPregnantBefore);
+            beneficiary.put("vblt_children", vbltChildren);
+            beneficiary.put("vblt_pregnant_or_breastfeeding", vbltPregnantOrBreastfeeding);
+            beneficiary.put("vblt_is_employed", vbltIsEmployed);
+            beneficiary.put("vblt_tested_hiv", vbltTestedHiv);
+            beneficiary.put("vblt_sexually_active", vbltSexuallyActive);
+            beneficiary.put("vblt_multiple_partners", vbltMultiplePartners);
+            beneficiary.put("vblt_is_migrant", vbltIsMigrant);
+            beneficiary.put("vblt_trafficking_victim", vbltTraffickingVictim);
+            beneficiary.put("vblt_sexual_exploitation", vbltSexualExploitation);
+            beneficiary.put("vblt_sexploitation_time", vbltSexploitationTime);
+            beneficiary.put("vblt_vbg_victim", vbltVbgVictim);
+            beneficiary.put("vblt_vbg_type", vbltVbgType);
+            beneficiary.put("vblt_vbg_time", vbltVbgTime);
+            beneficiary.put("vblt_alcohol_drugs_use", vbltAlcoholDrugsUse);
+            beneficiary.put("vblt_sti_history", vbltStiHistory);
+            beneficiary.put("vblt_sex_worker", vbltSexWorker);
+            beneficiary.put("vblt_house_sustainer", vbltHouseSustainer);
+        } else { // ensure online_id is updated first
+            beneficiary.put("online_id", id.toString());
+        }
+        return beneficiary;
+    }
+
+    public void update(BeneficiarySyncModel model, String timestamp) {
+        Long t = Long.valueOf(timestamp);
+        this.offlineId = model.getId();
+        this.dateUpdated = new Date(t);
+        this.nui = model.getNui();
+        this.name = model.getName();
+        this.surname = model.getSurname();
+        this.nickName = model.getNick_name();
+        this.partners.setId(model.getOrganization_id());
+        this.dateOfBirth = model.getDate_of_birth();
+        this.gender = model.getGender();
+        this.address = model.getAddress();
+        this.phoneNumber = model.getPhone_number();
+        this.EMail = model.getE_mail();
+        this.via = model.getVia();
+        this.partnerId=model.getPartner_id();
+        this.entryPoint = model.getEntry_point();
+        this.neighborhood.setId(model.getNeighbourhood_id());
+        this.us = new Us(model.getUs_id());
+        this.status = Integer.valueOf(model.getStatus());
+        this.vbltLivesWith = model.getVblt_lives_with();
+        this.vbltIsOrphan = (byte)model.getVblt_is_orphan();
+        this.vbltIsStudent = (byte)model.getVblt_is_student();
+        this.vbltSchoolGrade = model.getVblt_school_grade();
+        this.vbltSchoolName = model.getVblt_school_name();
+        this.vbltIsDeficient = (byte)model.getVblt_is_deficient();
+        this.vbltDeficiencyType = model.getVblt_deficiency_type();
+        this.vbltMarriedBefore = (byte)model.getVblt_married_before();
+        this.vbltPregnantBefore = (byte)model.getVblt_pregnant_before();
+        this.vbltChildren = (byte)model.getVblt_children();
+        this.vbltPregnantOrBreastfeeding = (byte)model.getVblt_pregnant_or_breastfeeding();
+        this.vbltIsEmployed = model.getVblt_is_employed();
+        this.vbltTestedHiv = model.getVblt_tested_hiv();
+        this.vbltSexuallyActive = (byte)model.getVblt_sexually_active();
+        this.vbltMultiplePartners = (byte)model.getVblt_multiple_partners();
+        this.vbltIsMigrant = (byte)model.getVblt_is_migrant();
+        this.vbltTraffickingVictim = (byte)model.getVblt_trafficking_victim();
+        this.vbltSexualExploitation = (byte)model.getVblt_sexual_exploitation();
+        this.vbltSexploitationTime = model.getVblt_sexploitation_time();
+        this.vbltVbgVictim = (byte)model.getVblt_vbg_victim();
+        this.vbltVbgType = model.getVblt_vbg_type();
+        this.vbltVbgTime = model.getVblt_vbg_time();
+        this.vbltAlcoholDrugsUse = (byte)model.getVblt_alcohol_drugs_use();
+        this.vbltStiHistory = (byte)model.getVblt_sti_history();
+        this.vbltSexWorker = (byte)model.getVblt_sex_worker();
+        this.vbltHouseSustainer = (byte)model.getVblt_house_sustainer();
+    }
 }
