+++ src/main/java/dlt/dltbackendmaster/domain/BeneficiariesInterventions.java
@@ -1,6 +1,9 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.Date;
 import javax.persistence.AttributeOverride;
 import javax.persistence.AttributeOverrides;
@@ -10,15 +13,36 @@ import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.NamedQueries;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+import dlt.dltbackendmaster.domain.watermelondb.BeneficiaryInterventionSyncModel;
+import dlt.dltbackendmaster.serializers.BeneficiarySerializer;
+import dlt.dltbackendmaster.serializers.LocalitySerializer;
+import dlt.dltbackendmaster.serializers.UsSerializer;
 
 /**
  * BeneficiariesInterventions generated by hbm2java
  */
 @Entity
 @Table(name = "beneficiaries_interventions", catalog = "dreams_db")
+@NamedQueries({ 
+	@NamedQuery(name = "BeneficiariesInterventions.findAll", query = "SELECT b FROM BeneficiariesInterventions b"),
+	@NamedQuery(name = "BeneficiariesInterventions.findByBeneficiaryAndSubService", 
+		query = "SELECT b FROM BeneficiariesInterventions b where b.beneficiaries.id = :beneficiary_id and b.subServices.id = :sub_service_id"),
+	@NamedQuery(name = "BeneficiariesInterventions.findByDateCreated", 
+		query = "select b from BeneficiariesInterventions b where b.dateUpdated is null and b.dateCreated > :lastpulledat"),
+	@NamedQuery(name = "BeneficiariesInterventions.findByDateUpdated",
+		query = "select b from BeneficiariesInterventions b where (b.dateUpdated >= :lastpulledat) or (b.dateUpdated >= :lastpulledat and b.dateCreated = b.dateUpdated)") })
 public class BeneficiariesInterventions implements java.io.Serializable {
 
 	private BeneficiariesInterventionsId id;
@@ -72,6 +96,27 @@ public class BeneficiariesInterventions implements java.io.Serializable {
 		this.dateUpdated = dateUpdated;
 		this.offlineId = offlineId;
 	}
+	
+	public BeneficiariesInterventions(BeneficiaryInterventionSyncModel model, String timestamp) throws ParseException {
+        Long t = Long.valueOf(timestamp);
+        Date regDate = new Date(t);
+        this.beneficiaries = new Beneficiaries(model.getBeneficiary_id());
+        this.subServices = new SubServices(model.getSub_service_id());
+        this.result = model.getResult();
+        DateFormat dateFormatter = new SimpleDateFormat("yyyy-MM-dd");
+        Date date1=dateFormatter.parse(model.getDate());  
+        this.id.setDate(date1);
+        this.us = new Us(model.getUs_id());
+        this.activistId = model.getActivist_id();
+        this.entryPoint = model.getEntry_point();
+        this.provider = model.getProvider();
+        this.remarks = model.getRemarks();
+        this.status = model.getStatus();
+        //this.offlineId = model.getBeneficiary_id() + "" + model.getSub_service_id();
+        this.offlineId = model.getId();
+        this.dateCreated = regDate;
+        this.dateUpdated = regDate;
+    }
 
 	@EmbeddedId
 
@@ -87,8 +132,11 @@ public class BeneficiariesInterventions implements java.io.Serializable {
 		this.id = id;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@JsonIgnore
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "beneficiary_id", nullable = false, insertable = false, updatable = false)
+	@JsonProperty("beneficiary")
+    @JsonSerialize(using = BeneficiarySerializer.class)
 	public Beneficiaries getBeneficiaries() {
 		return this.beneficiaries;
 	}
@@ -97,7 +145,8 @@ public class BeneficiariesInterventions implements java.io.Serializable {
 		this.beneficiaries = beneficiaries;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@JsonIgnore
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "sub_service_id", nullable = false, insertable = false, updatable = false)
 	public SubServices getSubServices() {
 		return this.subServices;
@@ -107,8 +156,10 @@ public class BeneficiariesInterventions implements java.io.Serializable {
 		this.subServices = subServices;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "us_id", nullable = false)
+	@JsonProperty("us")
+    @JsonSerialize(using = UsSerializer.class)
 	public Us getUs() {
 		return this.us;
 	}
@@ -217,5 +268,54 @@ public class BeneficiariesInterventions implements java.io.Serializable {
 	public void setOfflineId(String offlineId) {
 		this.offlineId = offlineId;
 	}
+	
+	public ObjectNode toObjectNode(String lastPulledAt) {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode beneficiaryIntervention = mapper.createObjectNode();
+        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
+        if (offlineId != null) {
+            beneficiaryIntervention.put("id", offlineId);
+        } else {
+            beneficiaryIntervention.put("id", id.toString());
+        }
+
+        if (dateUpdated == null || dateUpdated.after(dateCreated) || lastPulledAt == null || lastPulledAt.equals("null")) {
+            beneficiaryIntervention.put("beneficiary_id", beneficiaries.getId());
+            beneficiaryIntervention.put("sub_service_id", subServices.getId());
+            beneficiaryIntervention.put("result", result);
+            beneficiaryIntervention.put("date", dateFormat.format(id.getDate()));
+            beneficiaryIntervention.put("us_id", us.getId());
+            beneficiaryIntervention.put("result", result);
+            beneficiaryIntervention.put("activist_id", activistId);
+            beneficiaryIntervention.put("entry_point", entryPoint);
+            beneficiaryIntervention.put("provider", provider);
+            beneficiaryIntervention.put("remarks", remarks);
+            beneficiaryIntervention.put("status", status);
+            beneficiaryIntervention.put("online_id", id.toString()); // flag to control if entity is synchronized with the backend
+        } else { // ensure online_id is updated first
+            beneficiaryIntervention.put("online_id", id.toString());
+        }
+        return beneficiaryIntervention;
+    }
+	
+	public void update(BeneficiaryInterventionSyncModel model, String timestamp) throws ParseException {
+        Long t = Long.valueOf(timestamp);
+        
+        this.offlineId = model.getId();
+        this.dateUpdated = new Date(t);
+        this.beneficiaries.setId(model.getBeneficiary_id());
+        this.subServices.setId(model.getSub_service_id());
+        DateFormat dateFormatter = new SimpleDateFormat("yyyy-MM-dd");
+        Date date1=dateFormatter.parse(model.getDate());  
+        this.id.setDate(date1);;
+        this.us = new Us(model.getUs_id());
+        this.result = model.getResult();
+        this.activistId = model.getActivist_id();
+        this.entryPoint = model.getEntry_point();
+        this.provider = model.getProvider();
+        this.remarks = model.getRemarks();
+        this.status = model.getStatus();
+    }
 
 }
