+++ src/main/java/dlt/dltbackendmaster/domain/CurriculumsServices.java
@@ -1,95 +1,93 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
+import javax.persistence.AttributeOverride;
+import javax.persistence.AttributeOverrides;
 import javax.persistence.Column;
+import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 
 /**
- * Locality generated by hbm2java
+ * CurriculumsServices generated by hbm2java
  */
 @Entity
-@Table(name = "locality", catalog = "dreams_db")
-public class Locality implements java.io.Serializable {
+@Table(name = "curriculums_services", catalog = "dreams_db")
+public class CurriculumsServices implements java.io.Serializable {
 
-	private Integer id;
-	private District district;
-	private String name;
+	private CurriculumsServicesId id;
+	private Curriculums curriculums;
+	private Services services;
 	private String description;
 	private int status;
 	private int createdBy;
 	private Date dateCreated;
 	private Integer updatedBy;
 	private Date dateUpdated;
-	private Set<Neighborhood> neighborhoods = new HashSet<Neighborhood>(0);
-	private Set<Users> userses = new HashSet<Users>(0);
 
-	public Locality() {
+	public CurriculumsServices() {
 	}
 
-	public Locality(District district, String name, int status, int createdBy, Date dateCreated) {
-		this.district = district;
-		this.name = name;
+	public CurriculumsServices(CurriculumsServicesId id, Curriculums curriculums, Services services, int status,
+			int createdBy, Date dateCreated) {
+		this.id = id;
+		this.curriculums = curriculums;
+		this.services = services;
 		this.status = status;
 		this.createdBy = createdBy;
 		this.dateCreated = dateCreated;
 	}
 
-	public Locality(District district, String name, String description, int status, int createdBy, Date dateCreated,
-			Integer updatedBy, Date dateUpdated, Set<Neighborhood> neighborhoods, Set<Users> userses) {
-		this.district = district;
-		this.name = name;
+	public CurriculumsServices(CurriculumsServicesId id, Curriculums curriculums, Services services, String description,
+			int status, int createdBy, Date dateCreated, Integer updatedBy, Date dateUpdated) {
+		this.id = id;
+		this.curriculums = curriculums;
+		this.services = services;
 		this.description = description;
 		this.status = status;
 		this.createdBy = createdBy;
 		this.dateCreated = dateCreated;
 		this.updatedBy = updatedBy;
 		this.dateUpdated = dateUpdated;
-		this.neighborhoods = neighborhoods;
-		this.userses = userses;
 	}
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
+	@EmbeddedId
 
-	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
+	@AttributeOverrides({
+			@AttributeOverride(name = "curriculumId", column = @Column(name = "curriculum_id", nullable = false)),
+			@AttributeOverride(name = "serviceId", column = @Column(name = "service_id", nullable = false)) })
+	public CurriculumsServicesId getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(CurriculumsServicesId id) {
 		this.id = id;
 	}
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "district_id", nullable = false)
-	public District getDistrict() {
-		return this.district;
+	@JoinColumn(name = "curriculum_id", nullable = false, insertable = false, updatable = false)
+	public Curriculums getCurriculums() {
+		return this.curriculums;
 	}
 
-	public void setDistrict(District district) {
-		this.district = district;
+	public void setCurriculums(Curriculums curriculums) {
+		this.curriculums = curriculums;
 	}
 
-	@Column(name = "name", nullable = false, length = 150)
-	public String getName() {
-		return this.name;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "service_id", nullable = false, insertable = false, updatable = false)
+	public Services getServices() {
+		return this.services;
 	}
 
-	public void setName(String name) {
-		this.name = name;
+	public void setServices(Services services) {
+		this.services = services;
 	}
 
 	@Column(name = "description", length = 250)
@@ -146,24 +144,6 @@ public class Locality implements java.io.Serializable {
 
 	public void setDateUpdated(Date dateUpdated) {
 		this.dateUpdated = dateUpdated;
-	}
-
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "locality")
-	public Set<Neighborhood> getNeighborhoods() {
-		return this.neighborhoods;
-	}
-
-	public void setNeighborhoods(Set<Neighborhood> neighborhoods) {
-		this.neighborhoods = neighborhoods;
-	}
-
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "locality")
-	public Set<Users> getUserses() {
-		return this.userses;
-	}
-
-	public void setUserses(Set<Users> userses) {
-		this.userses = userses;
 	}
 
 }
