+++ src/main/java/dlt/dltbackendmaster/domain/Partners.java
@@ -1,5 +1,5 @@
 package dlt.dltbackendmaster.domain;
-// Generated Jan 25, 2022, 4:05:43 PM by Hibernate Tools 5.2.12.Final
+// Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
 import java.util.HashSet;
@@ -12,8 +12,8 @@ import static javax.persistence.GenerationType.IDENTITY;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
+import javax.persistence.NamedQueries;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
@@ -24,7 +24,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
 /**
- * Partner generated by hbm2java
+ * Partners generated by hbm2java
  */
 @Entity
 @Table(name = "partners", catalog = "dreams_db")
@@ -47,6 +47,7 @@ public class Partners implements java.io.Serializable {
 	private Integer updatedBy;
 	private Date dateUpdated;
 	private Set<Users> userses = new HashSet<Users>(0);
+	private Set<Beneficiaries> beneficiarieses = new HashSet<Beneficiaries>(0);
 
 	public Partners() {
 	}
@@ -61,7 +62,7 @@ public class Partners implements java.io.Serializable {
 
 	public Partners(District district, String name, String abbreviation, String description, String partnerType,
 			String logo, int status, int createdBy, Date dateCreated, Integer updatedBy, Date dateUpdated,
-			Set<Users> userses) {
+			Set<Users> userses, Set<Beneficiaries> beneficiarieses) {
 		this.district = district;
 		this.name = name;
 		this.abbreviation = abbreviation;
@@ -74,6 +75,7 @@ public class Partners implements java.io.Serializable {
 		this.updatedBy = updatedBy;
 		this.dateUpdated = dateUpdated;
 		this.userses = userses;
+		this.beneficiarieses = beneficiarieses;
 	}
 	
 	public Partners(Integer id) {
@@ -93,7 +95,7 @@ public class Partners implements java.io.Serializable {
 	}
 
 	@JsonIgnore
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "district_id")
 	public District getDistrict() {
 		return this.district;
@@ -196,7 +198,7 @@ public class Partners implements java.io.Serializable {
 	}
 
 	@JsonIgnore
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "partners")
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "partners")
 	public Set<Users> getUserses() {
 		return this.userses;
 	}
@@ -204,7 +206,17 @@ public class Partners implements java.io.Serializable {
 	public void setUserses(Set<Users> userses) {
 		this.userses = userses;
 	}
-	
+
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "partners")
+	public Set<Beneficiaries> getBeneficiarieses() {
+		return this.beneficiarieses;
+	}
+
+	public void setBeneficiarieses(Set<Beneficiaries> beneficiarieses) {
+		this.beneficiarieses = beneficiarieses;
+	}
+
 	public ObjectNode toObjectNode() {
 		ObjectMapper mapper = new ObjectMapper();
 		
@@ -217,5 +229,4 @@ public class Partners implements java.io.Serializable {
 		partner.put("online_id", id); // flag to control if entity is synchronized with the backend
 		return partner;
 	} 
-
 }
