+++ src/main/java/dlt/dltbackendmaster/domain/Province.java
@@ -1,5 +1,5 @@
 package dlt.dltbackendmaster.domain;
-// Generated Jan 25, 2022, 4:05:43 PM by Hibernate Tools 5.2.12.Final
+// Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
 import java.util.HashSet;
@@ -139,7 +139,7 @@ public class Province implements java.io.Serializable {
 	}
 
 	@JsonIgnore
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "province")
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "province")
 	public Set<District> getDistricts() {
 		return this.districts;
 	}
