+++ src/main/java/dlt/dltbackendmaster/domain/References.java
@@ -1,4 +1,4 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
@@ -16,6 +16,13 @@ import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+
+import dlt.dltbackendmaster.serializers.BeneficiarySerializer;
+import dlt.dltbackendmaster.serializers.UsersSerializer;
 
 /**
  * References generated by hbm2java
@@ -93,8 +100,10 @@ public class References implements java.io.Serializable {
 		this.id = id;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "beneficiary_id", nullable = false)
+	@JsonProperty("beneficiary")
+    @JsonSerialize(using = BeneficiarySerializer.class)
 	public Beneficiaries getBeneficiaries() {
 		return this.beneficiaries;
 	}
@@ -103,8 +112,10 @@ public class References implements java.io.Serializable {
 		this.beneficiaries = beneficiaries;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "notify_to", nullable = false)
+	@JsonProperty("notify_to_user")
+    @JsonSerialize(using = UsersSerializer.class)
 	public Users getUsers() {
 		return this.users;
 	}
@@ -250,7 +261,8 @@ public class References implements java.io.Serializable {
 		this.dateUpdated = dateUpdated;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "references")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "references")
 	public Set<ReferencesServices> getReferencesServiceses() {
 		return this.referencesServiceses;
 	}
