+++ src/main/java/dlt/dltbackendmaster/domain/ReferencesServices.java
@@ -1,102 +1,93 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
+import javax.persistence.AttributeOverride;
+import javax.persistence.AttributeOverrides;
 import javax.persistence.Column;
+import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 
 /**
- * UsType generated by hbm2java
+ * ReferencesServices generated by hbm2java
  */
 @Entity
-@Table(name = "us_type", catalog = "dreams_db")
-public class UsType implements java.io.Serializable {
+@Table(name = "references_services", catalog = "dreams_db")
+public class ReferencesServices implements java.io.Serializable {
 
-	private Integer id;
-	private String name;
-	private String level;
-	private String type;
+	private ReferencesServicesId id;
+	private References references;
+	private Services services;
 	private String description;
 	private int status;
 	private int createdBy;
 	private Date dateCreated;
 	private Integer updatedBy;
 	private Date dateUpdated;
-	private Set<Us> uses = new HashSet<Us>(0);
 
-	public UsType() {
+	public ReferencesServices() {
 	}
 
-	public UsType(String name, String level, String type, int status, int createdBy, Date dateCreated) {
-		this.name = name;
-		this.level = level;
-		this.type = type;
+	public ReferencesServices(ReferencesServicesId id, References references, Services services, int status,
+			int createdBy, Date dateCreated) {
+		this.id = id;
+		this.references = references;
+		this.services = services;
 		this.status = status;
 		this.createdBy = createdBy;
 		this.dateCreated = dateCreated;
 	}
 
-	public UsType(String name, String level, String type, String description, int status, int createdBy,
-			Date dateCreated, Integer updatedBy, Date dateUpdated, Set<Us> uses) {
-		this.name = name;
-		this.level = level;
-		this.type = type;
+	public ReferencesServices(ReferencesServicesId id, References references, Services services, String description,
+			int status, int createdBy, Date dateCreated, Integer updatedBy, Date dateUpdated) {
+		this.id = id;
+		this.references = references;
+		this.services = services;
 		this.description = description;
 		this.status = status;
 		this.createdBy = createdBy;
 		this.dateCreated = dateCreated;
 		this.updatedBy = updatedBy;
 		this.dateUpdated = dateUpdated;
-		this.uses = uses;
 	}
 
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
+	@EmbeddedId
 
-	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
+	@AttributeOverrides({
+			@AttributeOverride(name = "referenceId", column = @Column(name = "reference_id", nullable = false)),
+			@AttributeOverride(name = "serviceId", column = @Column(name = "service_id", nullable = false)) })
+	public ReferencesServicesId getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(ReferencesServicesId id) {
 		this.id = id;
 	}
 
-	@Column(name = "name", nullable = false, length = 50)
-	public String getName() {
-		return this.name;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "reference_id", nullable = false, insertable = false, updatable = false)
+	public References getReferences() {
+		return this.references;
 	}
 
-	public void setName(String name) {
-		this.name = name;
+	public void setReferences(References references) {
+		this.references = references;
 	}
 
-	@Column(name = "level", nullable = false, length = 50)
-	public String getLevel() {
-		return this.level;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "service_id", nullable = false, insertable = false, updatable = false)
+	public Services getServices() {
+		return this.services;
 	}
 
-	public void setLevel(String level) {
-		this.level = level;
-	}
-
-	@Column(name = "type", nullable = false, length = 150)
-	public String getType() {
-		return this.type;
-	}
-
-	public void setType(String type) {
-		this.type = type;
+	public void setServices(Services services) {
+		this.services = services;
 	}
 
 	@Column(name = "description", length = 250)
@@ -153,15 +144,6 @@ public class UsType implements java.io.Serializable {
 
 	public void setDateUpdated(Date dateUpdated) {
 		this.dateUpdated = dateUpdated;
-	}
-
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "usType")
-	public Set<Us> getUses() {
-		return this.uses;
-	}
-
-	public void setUses(Set<Us> uses) {
-		this.uses = uses;
 	}
 
 }
