+++ src/main/java/dlt/dltbackendmaster/domain/Services.java
@@ -1,4 +1,4 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
@@ -10,16 +10,29 @@ import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import static javax.persistence.GenerationType.IDENTITY;
 import javax.persistence.Id;
+import javax.persistence.NamedQuery;
+import javax.persistence.NamedQueries;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 
 /**
  * Services generated by hbm2java
  */
 @Entity
 @Table(name = "services", catalog = "dreams_db")
+@NamedQueries({ @NamedQuery(name = "Services.findAll", query = "SELECT c FROM Services c"),
+    @NamedQuery(name = "Services.findByDateCreated",
+                query = "SELECT c FROM Services c WHERE c.dateCreated = :lastpulledat"),
+    @NamedQuery(name = "Services.findByDateUpdated",
+                query = "SELECT c FROM Services c WHERE c.dateUpdated = :lastpulledat"),
+    @NamedQuery(name = "Services.findByServiceType",
+                query = "SELECT c FROM Services c WHERE c.serviceType = :serviceType") })
 public class Services implements java.io.Serializable {
 
 	private Integer id;
@@ -181,7 +194,8 @@ public class Services implements java.io.Serializable {
 		this.dateUpdated = dateUpdated;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "services")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "services")
 	public Set<ReferencesServices> getReferencesServiceses() {
 		return this.referencesServiceses;
 	}
@@ -190,7 +204,7 @@ public class Services implements java.io.Serializable {
 		this.referencesServiceses = referencesServiceses;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "services")
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "services")
 	public Set<SubServices> getSubServiceses() {
 		return this.subServiceses;
 	}
@@ -199,7 +213,8 @@ public class Services implements java.io.Serializable {
 		this.subServiceses = subServiceses;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "services")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "services")
 	public Set<CurriculumsServices> getCurriculumsServiceses() {
 		return this.curriculumsServiceses;
 	}
@@ -207,5 +222,20 @@ public class Services implements java.io.Serializable {
 	public void setCurriculumsServiceses(Set<CurriculumsServices> curriculumsServiceses) {
 		this.curriculumsServiceses = curriculumsServiceses;
 	}
+	
+	public ObjectNode toObjectNode() {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode service = mapper.createObjectNode();
+        service.put("id", id);
+        service.put("service_type", serviceType);
+        service.put("name", name);
+        service.put("description", description);
+        service.put("is_core_service", coreService);
+        service.put("is_hidden", hidden);
+        service.put("age_bands", ageBands);
+        service.put("status", status);
+        service.put("online_id", id); // flag to control if entity is synchronized with the backend
+        return service;
+    }
 
 }
