+++ src/main/java/dlt/dltbackendmaster/domain/SubServices.java
@@ -1,4 +1,4 @@
-package domains;
+package dlt.dltbackendmaster.domain;
 // Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
 
 import java.util.Date;
@@ -12,16 +12,33 @@ import static javax.persistence.GenerationType.IDENTITY;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.NamedQuery;
+import javax.persistence.NamedQueries;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+import dlt.dltbackendmaster.serializers.ServiceSerializer;
 
 /**
  * SubServices generated by hbm2java
  */
 @Entity
 @Table(name = "sub_services", catalog = "dreams_db")
+@NamedQueries({ @NamedQuery(name = "SubServices.findAll", query = "SELECT c FROM SubServices c"),
+    @NamedQuery(name = "SubServices.findByDateCreated",
+                query = "SELECT c FROM SubServices c WHERE c.dateCreated = :lastpulledat"),
+    @NamedQuery(name = "SubServices.findByDateUpdated",
+                query = "SELECT c FROM SubServices c WHERE c.dateUpdated = :lastpulledat"),
+    @NamedQuery(name = "SubServices.findByService",
+                query = "SELECT c FROM SubServices c WHERE c.services.id = :serviceId") })
 public class SubServices implements java.io.Serializable {
 
 	private Integer id;
@@ -65,6 +82,10 @@ public class SubServices implements java.io.Serializable {
 		this.dateUpdated = dateUpdated;
 		this.beneficiariesInterventionses = beneficiariesInterventionses;
 	}
+	
+	public SubServices(Integer id) {
+        this.id = id;
+    }
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
@@ -78,8 +99,10 @@ public class SubServices implements java.io.Serializable {
 		this.id = id;
 	}
 
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne(fetch = FetchType.EAGER)
 	@JoinColumn(name = "service_id", nullable = false)
+	@JsonProperty("service")
+    @JsonSerialize(using = ServiceSerializer.class)
 	public Services getServices() {
 		return this.services;
 	}
@@ -180,7 +203,8 @@ public class SubServices implements java.io.Serializable {
 		this.dateUpdated = dateUpdated;
 	}
 
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "subServices")
+	@JsonIgnore
+	@OneToMany(fetch = FetchType.EAGER, mappedBy = "subServices")
 	public Set<BeneficiariesInterventions> getBeneficiariesInterventionses() {
 		return this.beneficiariesInterventionses;
 	}
@@ -189,4 +213,18 @@ public class SubServices implements java.io.Serializable {
 		this.beneficiariesInterventionses = beneficiariesInterventionses;
 	}
 
+	public ObjectNode toObjectNode() {
+        ObjectMapper mapper = new ObjectMapper();
+        ObjectNode subService = mapper.createObjectNode();
+        subService.put("id", id);
+        subService.put("name", name);
+        subService.put("service_id", services.getId());
+        subService.put("is_mandatory", mandatory);
+        subService.put("is_hidden", hidden);
+        subService.put("remarks", remarks);
+        subService.put("sort_order", sortOrder);
+        subService.put("status", status);
+        subService.put("online_id", id); // flag to control if entity is synchronized with the backend
+        return subService;
+    }
 }
