--- src/main/java/dlt/dltbackendmaster/domain/domains/Profiles.java
@@ -1,129 +1,0 @@
-package domains;
-// Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
-
-import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-/**
- * Profiles generated by hbm2java
- */
-@Entity
-@Table(name = "profiles", catalog = "dreams_db")
-public class Profiles implements java.io.Serializable {
-
-	private Integer id;
-	private String name;
-	private String description;
-	private Date dateCreated;
-	private Integer createdBy;
-	private Date dateUpdated;
-	private Integer updatedBy;
-	private Set<Users> userses = new HashSet<Users>(0);
-
-	public Profiles() {
-	}
-
-	public Profiles(String name) {
-		this.name = name;
-	}
-
-	public Profiles(String name, String description, Date dateCreated, Integer createdBy, Date dateUpdated,
-			Integer updatedBy, Set<Users> userses) {
-		this.name = name;
-		this.description = description;
-		this.dateCreated = dateCreated;
-		this.createdBy = createdBy;
-		this.dateUpdated = dateUpdated;
-		this.updatedBy = updatedBy;
-		this.userses = userses;
-	}
-
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-
-	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
-		return this.id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	@Column(name = "name", nullable = false, length = 150)
-	public String getName() {
-		return this.name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	@Column(name = "description", length = 250)
-	public String getDescription() {
-		return this.description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_created", length = 19)
-	public Date getDateCreated() {
-		return this.dateCreated;
-	}
-
-	public void setDateCreated(Date dateCreated) {
-		this.dateCreated = dateCreated;
-	}
-
-	@Column(name = "created_by")
-	public Integer getCreatedBy() {
-		return this.createdBy;
-	}
-
-	public void setCreatedBy(Integer createdBy) {
-		this.createdBy = createdBy;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_updated", length = 19)
-	public Date getDateUpdated() {
-		return this.dateUpdated;
-	}
-
-	public void setDateUpdated(Date dateUpdated) {
-		this.dateUpdated = dateUpdated;
-	}
-
-	@Column(name = "updated_by")
-	public Integer getUpdatedBy() {
-		return this.updatedBy;
-	}
-
-	public void setUpdatedBy(Integer updatedBy) {
-		this.updatedBy = updatedBy;
-	}
-
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "profiles")
-	public Set<Users> getUserses() {
-		return this.userses;
-	}
-
-	public void setUserses(Set<Users> userses) {
-		this.userses = userses;
-	}
-
-}
