--- src/main/java/dlt/dltbackendmaster/domain/domains/Users.java
@@ -1,313 +1,0 @@
-package domains;
-// Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
-
-import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-/**
- * Users generated by hbm2java
- */
-@Entity
-@Table(name = "users", catalog = "dreams_db")
-public class Users implements java.io.Serializable {
-
-	private Integer id;
-	private Locality locality;
-	private Partners partners;
-	private Profiles profiles;
-	private Us us;
-	private String surname;
-	private String name;
-	private String phoneNumber;
-	private String email;
-	private String username;
-	private String password;
-	private String entryPoint;
-	private int status;
-	private Byte isLocked;
-	private Byte isExpired;
-	private Byte isCredentialsExpired;
-	private Byte isEnabled;
-	private int createdBy;
-	private Date dateCreated;
-	private Integer updatedBy;
-	private Date dateUpdated;
-	private String offlineId;
-	private Set<References> referenceses = new HashSet<References>(0);
-
-	public Users() {
-	}
-
-	public Users(Partners partners, Profiles profiles, String surname, String name, String phoneNumber, String email,
-			String username, String password, int status, int createdBy, Date dateCreated) {
-		this.partners = partners;
-		this.profiles = profiles;
-		this.surname = surname;
-		this.name = name;
-		this.phoneNumber = phoneNumber;
-		this.email = email;
-		this.username = username;
-		this.password = password;
-		this.status = status;
-		this.createdBy = createdBy;
-		this.dateCreated = dateCreated;
-	}
-
-	public Users(Locality locality, Partners partners, Profiles profiles, Us us, String surname, String name,
-			String phoneNumber, String email, String username, String password, String entryPoint, int status,
-			Byte isLocked, Byte isExpired, Byte isCredentialsExpired, Byte isEnabled, int createdBy, Date dateCreated,
-			Integer updatedBy, Date dateUpdated, String offlineId, Set<References> referenceses) {
-		this.locality = locality;
-		this.partners = partners;
-		this.profiles = profiles;
-		this.us = us;
-		this.surname = surname;
-		this.name = name;
-		this.phoneNumber = phoneNumber;
-		this.email = email;
-		this.username = username;
-		this.password = password;
-		this.entryPoint = entryPoint;
-		this.status = status;
-		this.isLocked = isLocked;
-		this.isExpired = isExpired;
-		this.isCredentialsExpired = isCredentialsExpired;
-		this.isEnabled = isEnabled;
-		this.createdBy = createdBy;
-		this.dateCreated = dateCreated;
-		this.updatedBy = updatedBy;
-		this.dateUpdated = dateUpdated;
-		this.offlineId = offlineId;
-		this.referenceses = referenceses;
-	}
-
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-
-	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
-		return this.id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "locality_id")
-	public Locality getLocality() {
-		return this.locality;
-	}
-
-	public void setLocality(Locality locality) {
-		this.locality = locality;
-	}
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "partner_id", nullable = false)
-	public Partners getPartners() {
-		return this.partners;
-	}
-
-	public void setPartners(Partners partners) {
-		this.partners = partners;
-	}
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "profile_id", nullable = false)
-	public Profiles getProfiles() {
-		return this.profiles;
-	}
-
-	public void setProfiles(Profiles profiles) {
-		this.profiles = profiles;
-	}
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "us_id")
-	public Us getUs() {
-		return this.us;
-	}
-
-	public void setUs(Us us) {
-		this.us = us;
-	}
-
-	@Column(name = "surname", nullable = false, length = 50)
-	public String getSurname() {
-		return this.surname;
-	}
-
-	public void setSurname(String surname) {
-		this.surname = surname;
-	}
-
-	@Column(name = "name", nullable = false, length = 150)
-	public String getName() {
-		return this.name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	@Column(name = "phone_number", nullable = false, length = 50)
-	public String getPhoneNumber() {
-		return this.phoneNumber;
-	}
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	@Column(name = "email", nullable = false, length = 150)
-	public String getEmail() {
-		return this.email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	@Column(name = "username", nullable = false, length = 50)
-	public String getUsername() {
-		return this.username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	@Column(name = "password", nullable = false, length = 150)
-	public String getPassword() {
-		return this.password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	@Column(name = "entry_point", length = 50)
-	public String getEntryPoint() {
-		return this.entryPoint;
-	}
-
-	public void setEntryPoint(String entryPoint) {
-		this.entryPoint = entryPoint;
-	}
-
-	@Column(name = "status", nullable = false)
-	public int getStatus() {
-		return this.status;
-	}
-
-	public void setStatus(int status) {
-		this.status = status;
-	}
-
-	@Column(name = "is_locked")
-	public Byte getIsLocked() {
-		return this.isLocked;
-	}
-
-	public void setIsLocked(Byte isLocked) {
-		this.isLocked = isLocked;
-	}
-
-	@Column(name = "is_expired")
-	public Byte getIsExpired() {
-		return this.isExpired;
-	}
-
-	public void setIsExpired(Byte isExpired) {
-		this.isExpired = isExpired;
-	}
-
-	@Column(name = "is_credentials_expired")
-	public Byte getIsCredentialsExpired() {
-		return this.isCredentialsExpired;
-	}
-
-	public void setIsCredentialsExpired(Byte isCredentialsExpired) {
-		this.isCredentialsExpired = isCredentialsExpired;
-	}
-
-	@Column(name = "is_enabled")
-	public Byte getIsEnabled() {
-		return this.isEnabled;
-	}
-
-	public void setIsEnabled(Byte isEnabled) {
-		this.isEnabled = isEnabled;
-	}
-
-	@Column(name = "created_by", nullable = false)
-	public int getCreatedBy() {
-		return this.createdBy;
-	}
-
-	public void setCreatedBy(int createdBy) {
-		this.createdBy = createdBy;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_created", nullable = false, length = 19)
-	public Date getDateCreated() {
-		return this.dateCreated;
-	}
-
-	public void setDateCreated(Date dateCreated) {
-		this.dateCreated = dateCreated;
-	}
-
-	@Column(name = "updated_by")
-	public Integer getUpdatedBy() {
-		return this.updatedBy;
-	}
-
-	public void setUpdatedBy(Integer updatedBy) {
-		this.updatedBy = updatedBy;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_updated", length = 19)
-	public Date getDateUpdated() {
-		return this.dateUpdated;
-	}
-
-	public void setDateUpdated(Date dateUpdated) {
-		this.dateUpdated = dateUpdated;
-	}
-
-	@Column(name = "offline_id", length = 45)
-	public String getOfflineId() {
-		return this.offlineId;
-	}
-
-	public void setOfflineId(String offlineId) {
-		this.offlineId = offlineId;
-	}
-
-	@OneToMany(fetch = FetchType.LAZY, mappedBy = "users")
-	public Set<References> getReferenceses() {
-		return this.referenceses;
-	}
-
-	public void setReferenceses(Set<References> referenceses) {
-		this.referenceses = referenceses;
-	}
-
-}
