--- src/main/java/dlt/dltbackendmaster/domain/domains/VulnerabilityHistory.java
@@ -1,159 +1,0 @@
-package domains;
-// Generated Jun 13, 2022, 9:37:47 AM by Hibernate Tools 5.2.12.Final
-
-import java.util.Date;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import static javax.persistence.GenerationType.IDENTITY;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-
-/**
- * VulnerabilityHistory generated by hbm2java
- */
-@Entity
-@Table(name = "vulnerability_history", catalog = "dreams_db")
-public class VulnerabilityHistory implements java.io.Serializable {
-
-	private Integer id;
-	private Beneficiaries beneficiaries;
-	private String vulnerability;
-	private String value;
-	private Date evaluationDate;
-	private Integer status;
-	private String remarks;
-	private String createdBy;
-	private Date dateCreated;
-	private String updatedBy;
-	private Date dateUpdated;
-
-	public VulnerabilityHistory() {
-	}
-
-	public VulnerabilityHistory(Beneficiaries beneficiaries, String vulnerability, String value, Date evaluationDate,
-			Integer status, String remarks, String createdBy, Date dateCreated, String updatedBy, Date dateUpdated) {
-		this.beneficiaries = beneficiaries;
-		this.vulnerability = vulnerability;
-		this.value = value;
-		this.evaluationDate = evaluationDate;
-		this.status = status;
-		this.remarks = remarks;
-		this.createdBy = createdBy;
-		this.dateCreated = dateCreated;
-		this.updatedBy = updatedBy;
-		this.dateUpdated = dateUpdated;
-	}
-
-	@Id
-	@GeneratedValue(strategy = IDENTITY)
-
-	@Column(name = "id", unique = true, nullable = false)
-	public Integer getId() {
-		return this.id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "beneficiary_id")
-	public Beneficiaries getBeneficiaries() {
-		return this.beneficiaries;
-	}
-
-	public void setBeneficiaries(Beneficiaries beneficiaries) {
-		this.beneficiaries = beneficiaries;
-	}
-
-	@Column(name = "vulnerability", length = 45)
-	public String getVulnerability() {
-		return this.vulnerability;
-	}
-
-	public void setVulnerability(String vulnerability) {
-		this.vulnerability = vulnerability;
-	}
-
-	@Column(name = "value", length = 45)
-	public String getValue() {
-		return this.value;
-	}
-
-	public void setValue(String value) {
-		this.value = value;
-	}
-
-	@Temporal(TemporalType.DATE)
-	@Column(name = "evaluation_date", length = 10)
-	public Date getEvaluationDate() {
-		return this.evaluationDate;
-	}
-
-	public void setEvaluationDate(Date evaluationDate) {
-		this.evaluationDate = evaluationDate;
-	}
-
-	@Column(name = "status")
-	public Integer getStatus() {
-		return this.status;
-	}
-
-	public void setStatus(Integer status) {
-		this.status = status;
-	}
-
-	@Column(name = "remarks", length = 45)
-	public String getRemarks() {
-		return this.remarks;
-	}
-
-	public void setRemarks(String remarks) {
-		this.remarks = remarks;
-	}
-
-	@Column(name = "created_by", length = 45)
-	public String getCreatedBy() {
-		return this.createdBy;
-	}
-
-	public void setCreatedBy(String createdBy) {
-		this.createdBy = createdBy;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_created", length = 19)
-	public Date getDateCreated() {
-		return this.dateCreated;
-	}
-
-	public void setDateCreated(Date dateCreated) {
-		this.dateCreated = dateCreated;
-	}
-
-	@Column(name = "updated_by", length = 45)
-	public String getUpdatedBy() {
-		return this.updatedBy;
-	}
-
-	public void setUpdatedBy(String updatedBy) {
-		this.updatedBy = updatedBy;
-	}
-
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "date_updated", length = 19)
-	public Date getDateUpdated() {
-		return this.dateUpdated;
-	}
-
-	public void setDateUpdated(Date dateUpdated) {
-		this.dateUpdated = dateUpdated;
-	}
-
-}
