+++ src/main/java/dlt/dltbackendmaster/domain/watermelondb/BeneficiarySyncModel.java
@@ -2,9 +2,6 @@ package dlt.dltbackendmaster.domain.watermelondb;
 
 import java.util.Date;
 
-import dlt.dltbackendmaster.domain.DeficiencyType;
-import dlt.dltbackendmaster.domain.LivesWith;
-
 public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 {
     private String nui;
@@ -16,37 +13,37 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
     private String address;
     private String phone_number;
     private String e_mail;
-    private Boolean via;
+    private int via;
     private Integer partner_id;
     private String entry_point;
     private Integer neighbourhood_id;
     private Integer us_id;
     private String vblt_lives_with;
-    private Boolean vblt_is_orphan;
-    private Boolean vblt_is_student;
+    private int vblt_is_orphan;
+    private int vblt_is_student;
     private Integer vblt_school_grade;
     private String vblt_school_name;
-    private Boolean vblt_is_deficient;
+    private int vblt_is_deficient;
     private String vblt_deficiency_type;
-    private Boolean vblt_married_before;
-    private Boolean vblt_pregnant_before;
-    private Boolean vblt_children;
-    private Boolean vblt_pregnant_or_breastfeeding;
+    private int vblt_married_before;
+    private int vblt_pregnant_before;
+    private int vblt_children;
+    private int vblt_pregnant_or_breastfeeding;
     private String vblt_is_employed;
     private String vblt_tested_hiv;
-    private Boolean vblt_sexually_active;
-    private Boolean vblt_multiple_partners;
-    private Boolean vblt_is_migrant;
-    private Boolean vblt_trafficking_victim;
-    private Boolean vblt_sexual_exploitation;
+    private int vblt_sexually_active;
+    private int vblt_multiple_partners;
+    private int vblt_is_migrant;
+    private int vblt_trafficking_victim;
+    private int vblt_sexual_exploitation;
     private String vblt_sexploitation_time;
-    private Boolean vblt_vbg_victim;
+    private int vblt_vbg_victim;
     private String vblt_vbg_type;
     private String vblt_vbg_time;
-    private Boolean vblt_alcohol_drugs_use;
-    private Boolean vblt_sti_history;
-    private Boolean vblt_sex_worker;
-    private Boolean vblt_house_sustainer;
+    private int vblt_alcohol_drugs_use;
+    private int vblt_sti_history;
+    private int vblt_sex_worker;
+    private int vblt_house_sustainer;
 
     public BeneficiarySyncModel() {}
 
@@ -118,11 +115,11 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
         return e_mail;
     }
 
-    public Boolean getVia() {
+    public int getVia() {
         return via;
     }
 
-    public void setVia(Boolean via) {
+    public void setVia(int via) {
         this.via = via;
     }
 
@@ -166,27 +163,27 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_lives_with = vblt_lives_with;
 	}
 
-	public Boolean getVblt_is_orphan() {
+	public int getVblt_is_orphan() {
 		return vblt_is_orphan;
 	}
 
-	public void setVblt_is_orphan(Boolean vblt_is_orphan) {
+	public void setVblt_is_orphan(int vblt_is_orphan) {
 		this.vblt_is_orphan = vblt_is_orphan;
 	}
 
-	public Boolean getVblt_is_student() {
+	public int getVblt_is_student() {
 		return vblt_is_student;
 	}
 
-	public void setVblt_is_student(Boolean vblt_is_student) {
+	public void setVblt_is_student(int vblt_is_student) {
 		this.vblt_is_student = vblt_is_student;
 	}
 
-	public Integer getVblt_school_grade() {
+	public int getVblt_school_grade() {
 		return vblt_school_grade;
 	}
 
-	public void setVblt_school_grade(Integer vblt_school_grade) {
+	public void setVblt_school_grade(int vblt_school_grade) {
 		this.vblt_school_grade = vblt_school_grade;
 	}
 
@@ -198,11 +195,11 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_school_name = vblt_school_name;
 	}
 
-	public Boolean getVblt_is_deficient() {
+	public int getVblt_is_deficient() {
 		return vblt_is_deficient;
 	}
 
-	public void setVblt_is_deficient(Boolean vblt_is_deficient) {
+	public void setVblt_is_deficient(int vblt_is_deficient) {
 		this.vblt_is_deficient = vblt_is_deficient;
 	}
 
@@ -214,35 +211,35 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_deficiency_type = vblt_deficiency_type;
 	}
 
-	public Boolean getVblt_married_before() {
+	public int getVblt_married_before() {
 		return vblt_married_before;
 	}
 
-	public void setVblt_married_before(Boolean vblt_married_before) {
+	public void setVblt_married_before(int vblt_married_before) {
 		this.vblt_married_before = vblt_married_before;
 	}
 
-	public Boolean getVblt_pregnant_before() {
+	public int getVblt_pregnant_before() {
 		return vblt_pregnant_before;
 	}
 
-	public void setVblt_pregnant_before(Boolean vblt_pregnant_before) {
+	public void setVblt_pregnant_before(int vblt_pregnant_before) {
 		this.vblt_pregnant_before = vblt_pregnant_before;
 	}
 
-	public Boolean getVblt_children() {
+	public int getVblt_children() {
 		return vblt_children;
 	}
 
-	public void setVblt_children(Boolean vblt_children) {
+	public void setVblt_children(int vblt_children) {
 		this.vblt_children = vblt_children;
 	}
 
-	public Boolean getVblt_pregnant_or_breastfeeding() {
+	public int getVblt_pregnant_or_breastfeeding() {
 		return vblt_pregnant_or_breastfeeding;
 	}
 
-	public void setVblt_pregnant_or_breastfeeding(Boolean vblt_pregnant_or_breastfeeding) {
+	public void setVblt_pregnant_or_breastfeeding(int vblt_pregnant_or_breastfeeding) {
 		this.vblt_pregnant_or_breastfeeding = vblt_pregnant_or_breastfeeding;
 	}
 
@@ -262,43 +259,43 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_tested_hiv = vblt_tested_hiv;
 	}
 
-	public Boolean getVblt_sexually_active() {
+	public int getVblt_sexually_active() {
 		return vblt_sexually_active;
 	}
 
-	public void setVblt_sexually_active(Boolean vblt_sexually_active) {
+	public void setVblt_sexually_active(int vblt_sexually_active) {
 		this.vblt_sexually_active = vblt_sexually_active;
 	}
 
-	public Boolean getVblt_multiple_partners() {
+	public int getVblt_multiple_partners() {
 		return vblt_multiple_partners;
 	}
 
-	public void setVblt_multiple_partners(Boolean vblt_multiple_partners) {
+	public void setVblt_multiple_partners(int vblt_multiple_partners) {
 		this.vblt_multiple_partners = vblt_multiple_partners;
 	}
 
-	public Boolean getVblt_is_migrant() {
+	public int getVblt_is_migrant() {
 		return vblt_is_migrant;
 	}
 
-	public void setVblt_is_migrant(Boolean vblt_is_migrant) {
+	public void setVblt_is_migrant(int vblt_is_migrant) {
 		this.vblt_is_migrant = vblt_is_migrant;
 	}
 
-	public Boolean getVblt_trafficking_victim() {
+	public int getVblt_trafficking_victim() {
 		return vblt_trafficking_victim;
 	}
 
-	public void setVblt_trafficking_victim(Boolean vblt_trafficking_victim) {
+	public void setVblt_trafficking_victim(int vblt_trafficking_victim) {
 		this.vblt_trafficking_victim = vblt_trafficking_victim;
 	}
 
-	public Boolean getVblt_sexual_exploitation() {
+	public int getVblt_sexual_exploitation() {
 		return vblt_sexual_exploitation;
 	}
 
-	public void setVblt_sexual_exploitation(Boolean vblt_sexual_exploitation) {
+	public void setVblt_sexual_exploitation(int vblt_sexual_exploitation) {
 		this.vblt_sexual_exploitation = vblt_sexual_exploitation;
 	}
 
@@ -310,11 +307,11 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_sexploitation_time = vblt_sexploitation_time;
 	}
 
-	public Boolean getVblt_vbg_victim() {
+	public int getVblt_vbg_victim() {
 		return vblt_vbg_victim;
 	}
 
-	public void setVblt_vbg_victim(Boolean vblt_vbg_victim) {
+	public void setVblt_vbg_victim(int vblt_vbg_victim) {
 		this.vblt_vbg_victim = vblt_vbg_victim;
 	}
 
@@ -334,35 +331,35 @@ public class BeneficiarySyncModel extends BasicLifeCycleSyncModel
 		this.vblt_vbg_time = vblt_vbg_time;
 	}
 
-	public Boolean getVblt_alcohol_drugs_use() {
+	public int getVblt_alcohol_drugs_use() {
 		return vblt_alcohol_drugs_use;
 	}
 
-	public void setVblt_alcohol_drugs_use(Boolean vblt_alcohol_drugs_use) {
+	public void setVblt_alcohol_drugs_use(int vblt_alcohol_drugs_use) {
 		this.vblt_alcohol_drugs_use = vblt_alcohol_drugs_use;
 	}
 
-	public Boolean getVblt_sti_history() {
+	public int getVblt_sti_history() {
 		return vblt_sti_history;
 	}
 
-	public void setVblt_sti_history(Boolean vblt_sti_history) {
+	public void setVblt_sti_history(int vblt_sti_history) {
 		this.vblt_sti_history = vblt_sti_history;
 	}
 
-	public Boolean getVblt_sex_worker() {
+	public int getVblt_sex_worker() {
 		return vblt_sex_worker;
 	}
 
-	public void setVblt_sex_worker(Boolean vblt_sex_worker) {
+	public void setVblt_sex_worker(int vblt_sex_worker) {
 		this.vblt_sex_worker = vblt_sex_worker;
 	}
 
-	public Boolean getVblt_house_sustainer() {
+	public int getVblt_house_sustainer() {
 		return vblt_house_sustainer;
 	}
 
-	public void setVblt_house_sustainer(Boolean vblt_house_sustainer) {
+	public void setVblt_house_sustainer(int vblt_house_sustainer) {
 		this.vblt_house_sustainer = vblt_house_sustainer;
 	}
