+++ src/main/java/dlt/dltbackendmaster/serializers/ServiceSerializer.java
@@ -6,18 +6,18 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
-import dlt.dltbackendmaster.domain.Service;
+import dlt.dltbackendmaster.domain.Services;
 
-public class ServiceSerializer extends JsonSerializer<Service>
+public class ServiceSerializer extends JsonSerializer<Services>
 {
     @Override
-    public void serialize(Service value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
-        Service service = new Service();
+    public void serialize(Services value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+        Services service = new Services();
         service.setId(value.getId());
         service.setName(value.getName());
         service.setDescription(value.getDescription());
-        service.setIsHidden(value.getIsHidden());
-        service.setIsCoreService(service.getIsCoreService());
+        service.setHidden(value.getHidden());
+        service.setCoreService(service.getCoreService());
         service.setStatus(service.getStatus());
         service.setCreatedBy(value.getCreatedBy());
         service.setDateCreated(value.getDateCreated());
