+++ src/main/java/dlt/dltbackendmaster/serializers/UsTypeSerializer.java
@@ -19,10 +19,10 @@ public class UsTypeSerializer extends JsonSerializer<UsType> {
 		usType.setLevel(value.getLevel());
 		usType.setType(value.getType());
 		usType.setStatus(value.getStatus());
-		usType.setDateCreated(value.getDateCreated());
+		/*usType.setDateCreated(value.getDateCreated());
 		usType.setDateUpdated(value.getDateUpdated());
 		usType.setCreatedBy(value.getCreatedBy());
-		usType.setUpdatedBy(value.getUpdatedBy());
+		usType.setUpdatedBy(value.getUpdatedBy());*/
 		gen.writeObject(usType);
 	}
